improve codebase maintainability
