+++ src/components/ActionsMenu.tsx
@@ -1,7 +1,7 @@
 import { actionsMenu, loadingScreen, openInYtBtn } from "../lib/dom";
 import { addToCollection, createPlaylist } from "../lib/libraryUtils";
 import { $, notify } from "../lib/utils";
-import fetchList from "../scripts/fetchList";
+import fetchList from "../modules/fetchList";
 import { appendToQueuelist } from "../scripts/queue";
 import { store } from "../lib/store";
 import './ActionsMenu.css';
