+++ src/lib/utils.ts
@@ -6,7 +6,7 @@ import ListItem from "../components/ListItem";
 import StreamItem from "../components/StreamItem";
 import player from "./player";
 import { getSaved, store } from "./store";
-import fetchList from "../scripts/fetchList";
+import fetchList from "../modules/fetchList";
