+++ src/main.ts
@@ -7,4 +7,4 @@ import './scripts/list';
 import './scripts/audioEvents';
 import './scripts/library';
 import './components/Settings';
-import('./scripts/start').then(mod => mod.default());
+import('./modules/start').then(mod => mod.default());
