+++ src/modules/miniPlayer.ts
